+++ src/CommunityPatchLauncher/Tasks/DownloadCommunityPatchTask.cs
@@ -123,7 +123,7 @@ namespace CommunityPatchLauncher.Tasks
         /// <returns>The string to download the file from</returns>
         protected string GetDownloadString(string communityPatch)
         {
-            string downloadPath = GetSetting<string>(communityPatchType + "/" + "URI");
+            string downloadPath = GetSetting<string>(communityPatch + "/" + "URI");
             return downloadPath ?? string.Empty;
         }