+++ src/CommunityPatchLauncher/UserControls/ProgressBarDialog.xaml.cs
@@ -1,0 +1,33 @@
+﻿using CommunityPatchLauncher.UserControls.SpecialTypes;
+using CommunityPatchLauncher.ViewModels;
+using FontAwesome.WPF;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace CommunityPatchLauncher.UserControls
+{
+    /// <summary>
+    /// Interaction logic for ProgressBarDialog.xaml
+    /// </summary>
+    public partial class ProgressBarDialog : UserControl
+    {
+
+        public ProgressBarDialog()
+        {
+            InitializeComponent();
+            DataContext = new ProgressBarDialogViewModel();
+        }
+    }
+}